Team00/src/requirements.txt
altair==5.1.2
asttokens==2.4.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.1
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
exceptiongroup==1.1.3
executing==2.0.1
flake8==6.1.0
fonttools==4.43.1
idna==3.4
interchange==2021.0.4
ipython==8.16.1
ipython-cypher==0.2.6
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mccabe==0.7.0
monotonic==1.6
mypy==1.6.1
mypy-extensions==1.0.0
neo4j==5.13.0
neo4jrestclient==2.1.1
neo4jupyter==0.2.0
networkx==3.2.1
numpy==1.26.1
packaging==23.2
pandas==2.1.2
pandas-stubs==2.1.1.230928
pansi==2020.7.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
prettytable==3.9.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
py2neo==2021.2.4
pycodestyle==2.11.1
pyflakes==3.1.0
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.6
scipy==1.11.3
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
tomli==2.0.1
toolz==0.12.0
traitlets==5.12.0
types-beautifulsoup4==4.12.0.6
types-html5lib==1.1.11.15
types-Pillow==10.1.0.0
types-pytz==2023.3.1.1
types-requests==2.31.0.10
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.8
